compose: "RHCEPH" product short name
We recently changed our RHEL-based product's short name in the product
metadata to "RHCEPH", to align our composes with other internal tooling.

Unfortunately the Ubuntu product_short value is hardcoded in
rhcephcompose, so we need to change the code here in the rhcephcompose
application itself, instead of changing it in our configuration
metadata.