artifacts: handle "dbg" packages
Add a new property to BinaryArtifact objects, "dbg_parent". This will
let us identify binaries as "-dbg" or not, and provide us with the
binary's parent name.

The purpose of this change is to expose this data to callers (Compose
class). Eventually we can check this parent name against our comps
manifest. This will allow us to add each package's -dbg bits into a
compose without having to list each "foo-dbg" and "foo-dbg" package pair
explicitly in the comps .xml manifest.