compose: store -dbg packages in COMPOSE_ID-dbg
If a BuildArtifact is a -dbg package, and its parent package is listed
in a comps manifest, add that -dbg package to the compose.

The purpose of this change is to ship -dbg packages in the product
deliverable for customers.

Use a separate -dbg directory so we don't duplicate so much data. RHCS 2
is about 8GB when we copy the dbg packages into each variant, and only
2.3GB when we route everything to this single -dbg directory.

Add a "include_dbg" configuration option to control this -dbg dir
(useful for speeding up test composes prior to GA).

A future change could turn this -dbg directory into a proper repository.