verify cached files' checksums
Prior to this change, we would blindly trust the files in the cache were
good to use without validating the cached files' contents. This could
cause issues when rhcephcompose crashed and left the cache in an invalid
state, or some network interruption or other disk issue caused our cache
files to exist with invalid contents.

When parsing the data for each build artifact, store the artifacts'
sha512 checksum values from the chacra server, and validate that the
checksums match our cached files. If they do not match, exit with an
error.

The purpose of this change is to make it safer to trust the download
operations and cache integrity.