artifacts: sanity-check download() directory args
It's easy to accidentally write code that downloads artifacts to a
directory that doesn't yet exist. When that happens, download() will
quietly write the files to a *file* named by the "cache_dir" or
"dest_dir" arg, rather than writing those files into the respective
directories.

To guard against this in new code, sanity-check the directory args to
download() and ensure that they really are directories.