artifacts: cache local checksum operations
Prior to this change, we would calculate the sha512 checksum of a
cached file every time we used it. This was fast for small packages, but
takes several seconds for large -dbg packages.

Save an in-memory set of cache files we've verified so the
verify_checksum() call is less expensive.