compose: refactor with create_repo() method
Break out the final bits of the run_distro() method into a create_repo()
method that does the rest of the work.

Add (poor bare-bones) test coverage for this method so we exercise it
during pytest runs.

In addition to generating repos for each variant, eventually I plan to
use this method to generate a "dbg" repository outside of
self.output_dir.