compose: raise on wrong dists in build lists
Search each build list file for mentions of any other distros. Raise an
error if we find this.

For example, if we have two build lists:

  builds = {
    'trusty': 'builds-trusty.txt',
    'xenial': 'builds-xenial.txt',
  }

... ensure that "builds-trusty.txt" does not contain any builds with
"xenial" in the Name-Version-Relese, and vice-versa.

The purpose of this change is to avoid accidentally mixing up the build
lists and shipping the wrong packages for a distro.