.travis.yml: use older pytest and py libs on py26
Our Python 2.6 support is to support RH's rel-eng environment
(eng-rhel-6). This environment has some older versions of pytest and py.
When these versions don't line up with the newer versions we use in
Travis CI, it makes it harder to catch bugs until we integrate
internally.

Make our py26 libs in Travis CI match the ones we use internally.

This means we have to drop the flake8 run on py26... not a big deal,
since we run it on py27 and py3.