travisci: install newer py3 kobo
There is a new pull request in progress to add Python 3 support to Kobo.

Use that version instead of the old/stale python-3-dev branch.

Install it in the Python 2 tests as well for more coverage. I'm not too
concerned regarding eng-rhel-6 because we have a modern kobo version
there (0.5.2 as of today), so this is not too far off.