artifacts: fix cache_dir creation
Prior to this change, when we called download(cache_dir), we failed to
dynamically create the cache directory if it did not already exist. We
were calling makedirs() on a non-existant artifact object attribute, not
the method argument.

The reason we did not hit this code path in rhcephcompose until now is
that we have already called makedirs in the main Compose class, before
we get to the individual PackageArtifact classes.

In fact, this error appears to be a bad copy-and-paste from the Compose
class.