artifacts: support persistent requests session
Add a new optional "session" kwarg to download(), so we can pass in a
persistent requests.Session() object. This allows us to optimize a
large set of artifact downloads.

If we don't set "session" in the caller, just initialize a fresh
requests Session as before.

Note, no callers use this yet.